requirements.txt
Flask==3.0.2
PyGithub==2.2.0
python-dotenv==1.0.1
requests==2.31.0 